* Added CommandCollection interface to facilitate adding commands from class files (outside of CommandProcessor) that contain multiple command definitions. CommandProcessor does not have to be aware of what specific commands are contained in the individual CommandCollection implementations, so CommandCollection implementations may now be extended without modifying CommandProcessor.  * Fixed overflow in InventoryTools.getAmountOfItem. Now returns an int.  * Added !take command for taking an item from a player.  * Fixed bugs related to spawn points when exiting through the Free Market portal to the default map, Perion, when the saved location data is missing.  * Allow warp to a map if the WZs for a portal have an invalid destination.  * Less sensitive fast HP/MP regeneration detection.  * Removed infraction triggers in AP and SP allocation because latency or tubi enabled clients may fire them.  * Identified cause of Energy Charge effect showing in place of Dash.

+++ trunk/src/argonms/common/character/inventory/InventoryTools.java
@@ -445,9 +445,9 @@ public final class InventoryTools {
 			return removeFromInventory(p.getInventory(type), itemId, quantity);
 		} else {
 			Inventory inv = p.getInventory(type);
-			short start = getAmountOfItem(inv, itemId);
+			int start = getAmountOfItem(inv, itemId);
 			UpdatedSlots updatedSlots = removeFromInventory(inv, itemId, quantity);
-			short end = getAmountOfItem(inv, itemId);
+			int end = getAmountOfItem(inv, itemId);
 			quantity -= start - end;
 			if (quantity > 0)
 				updatedSlots.union(removeFromInventory(p.getInventory(InventoryType.EQUIPPED), itemId, quantity));
@@ -478,8 +478,8 @@ public final class InventoryTools {
 		return true;
 	}
 
-	public static short getAmountOfItem(Inventory inv, int itemId) {
-		short total = 0;
+	public static int getAmountOfItem(Inventory inv, int itemId) {
+		int total = 0;
 		for (Short slot : inv.getItemSlots(itemId))
 			total += inv.get(slot.shortValue()).getQuantity();
 		return total;
